# internal/openapi/openapi.yaml
openapi: "3.0.3"
info:
  title: FabAID Manager API
  description: NSF grant/project tracking — effort, budgets, subawards, invoices, and statements of work.
  version: "1.0.0"
  license:
    name: Apache-2.0
servers:
  - url: /api/v1
    description: API v1
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Session cookie or API key (Bearer fabaid_…)
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
    Grant:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        award_number:
          type: string
        title:
          type: string
        pi_name:
          type: string
        institution:
          type: string
        agency:
          type: string
        start_date:
          type: string
          format: date
        end_date:
          type: string
          format: date
        total_budget:
          type: number
          format: double
        salary_escalation_rate:
          type: number
          format: double
        status:
          type: string
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
...
  parameters:
    grantID:
      name: grantID
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    personnelID:
      name: personnelID
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
security:
  - bearerAuth: []
  - cookieAuth: []
paths:
...
  /grants:
    get:
      tags: [grants]
      summary: List all grants
      operationId: listGrants
      responses:
        "200":
          description: Array of grants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Grant"
    post:
      tags: [grants]
      summary: Create a grant
      operationId: createGrant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Grant"
      responses:
        "201":
          description: Created grant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Grant"
  /grants/{grantID}:
    parameters:
      - $ref: "#/components/parameters/grantID"
    get:
      tags: [grants]
      summary: Get a grant by ID
      operationId: getGrant
      responses:
        "200":
          description: Grant object
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Grant"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    put:
      tags: [grants]
      summary: Update a grant
      operationId: updateGrant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Grant"
      responses:
        "200":
          description: Updated grant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Grant"
    delete:
      tags: [grants]
      summary: Delete a grant
      operationId: deleteGrant
      responses:
        "204":
          description: Deleted

  # ── WBS Areas ───────────────────────────────────────────
  /grants/{grantID}/wbs:
    parameters:
      - $ref: "#/components/parameters/grantID"
    get:
      tags: [wbs]
      summary: List WBS areas for a grant
      operationId: listWBSAreas
      responses:
        "200":
          description: Array of WBS areas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WBSArea"
    post:
      tags: [wbs]
      summary: Create a WBS area
      operationId: createWBSArea
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WBSArea"
      responses:
        "201":
          description: Created WBS area
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WBSArea"
  /grants/{grantID}/wbs/effort-summary:
    parameters:
      - $ref: "#/components/parameters/grantID"
    get:
      tags: [wbs]
      summary: Get WBS effort summary across fiscal years
      operationId: wbsEffortSummary
      responses:
        "200":
          description: Array of effort summaries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WBSEffortSummary"
  /grants/{grantID}/wbs/{wbsID}:
    parameters:
      - $ref: "#/components/parameters/grantID"
      - name: wbsID
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags: [wbs]
      summary: Get a WBS area
      operationId: getWBSArea
      responses:
        "200":
          description: WBS area
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WBSArea"
    put:
      tags: [wbs]
      summary: Update a WBS area
      operationId: updateWBSArea
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WBSArea"
      responses:
        "200":
          description: Updated WBS area
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WBSArea"
    delete:
      tags: [wbs]
      summary: Delete a WBS area
      operationId: deleteWBSArea
      responses:
        "204":
          description: Deleted

  # ── Personnel ───────────────────────────────────────────
  /grants/{grantID}/personnel:
    parameters:
      - $ref: "#/components/parameters/grantID"
    get:
      tags: [personnel]
      summary: List personnel for a grant
      operationId: listPersonnel
      responses:
        "200":
          description: Array of personnel
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Personnel"
    post:
      tags: [personnel]
      summary: Create a personnel record
      operationId: createPersonnel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Personnel"
      responses:
        "201":
          description: Created personnel
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Personnel"
  /grants/{grantID}/personnel/titles:
    parameters:
      - $ref: "#/components/parameters/grantID"
    get:
      tags: [personnel]
      summary: List distinct personnel titles
      operationId: listPersonnelTitles
      responses:
        "200":
          description: Array of title strings
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
  /grants/{grantID}/personnel/{personnelID}:
    parameters:
      - $ref: "#/components/parameters/grantID"
      - $ref: "#/components/parameters/personnelID"
    put:
      tags: [personnel]
      summary: Update a personnel record
      operationId: updatePersonnel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Personnel"
      responses:
        "200":
          description: Updated personnel
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Personnel"
    delete:
      tags: [personnel]
      summary: Delete a personnel record
      operationId: deletePersonnel
      responses:
        "204":
          description: Deleted
  /grants/{grantID}/personnel/{personnelID}/budget-summary:
    parameters:
      - $ref: "#/components/parameters/grantID"
      - $ref: "#/components/parameters/personnelID"
    get:
      tags: [personnel]
      summary: Budget summary for a person across years
      operationId: personnelBudgetSummary
      responses:
        "200":
          description: Array of budget entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PersonnelBudgetEntry"
  /grants/{grantID}/personnel/{personnelID}/default-wbs:
    parameters:
      - $ref: "#/components/parameters/grantID"
      - $ref: "#/components/parameters/personnelID"
    get:
      tags: [personnel]
      summary: List default WBS allocations for a person
      operationId: listPersonnelDefaultWBS
      responses:
        "200":
          description: Array of WBS allocations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PersonnelDefaultWBS"
    put:
      tags: [personnel]
      summary: Set default WBS allocations for a person
      operationId: setPersonnelDefaultWBS
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/PersonnelDefaultWBS"
      responses:
        "200":
          description: Saved allocations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PersonnelDefaultWBS"
  # ── Subawards ───────────────────────────────────────────
  /grants/{grantID}/subawards:
    parameters:
      - $ref: "#/components/parameters/grantID"
    get:
      tags: [subawards]
      summary: List subawards for a grant
      operationId: listSubawards
      responses:
        "200":
          description: Array of subawards
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Subaward"
    post:
      tags: [subawards]
      summary: Create a subaward
      operationId: createSubaward
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Subaward"
      responses:
        "201":
          description: Created subaward
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subaward"
  /grants/{grantID}/subawards/{subawardID}:
    parameters:
      - $ref: "#/components/parameters/grantID"
      - $ref: "#/components/parameters/subawardID"
    put:
      tags: [subawards]
      summary: Update a subaward
      operationId: updateSubaward
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Subaward"
      responses:
        "200":
          description: Updated subaward
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subaward"
    delete:
      tags: [subawards]
      summary: Delete a subaward
      operationId: deleteSubaward
      responses:
        "204":
          description: Deleted

  # ── Invoices ────────────────────────────────────────────
  /grants/{grantID}/subawards/{subawardID}/invoices:
    parameters:
      - $ref: "#/components/parameters/grantID"
      - $ref: "#/components/parameters/subawardID"
    get:
      tags: [invoices]
      summary: List invoices for a subaward
      operationId: listInvoices
      responses:
        "200":
          description: Array of invoices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Invoice"
    post:
      tags: [invoices]
      summary: Create an invoice
      operationId: createInvoice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Invoice"
      responses:
        "201":
          description: Created invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
  /grants/{grantID}/subawards/{subawardID}/invoices/{invoiceID}/status:
    parameters:
      - $ref: "#/components/parameters/grantID"
      - $ref: "#/components/parameters/subawardID"
      - name: invoiceID
        in: path
        required: true
        schema:
          type: string
          format: uuid
    patch:
      tags: [invoices]
      summary: Update invoice status
      operationId: updateInvoiceStatus
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
              required: [status]
      responses:
        "200":
          description: Updated status
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
  /grants/{grantID}/subawards/{subawardID}/invoices/{invoiceID}/upload:
    parameters:
      - $ref: "#/components/parameters/grantID"
      - $ref: "#/components/parameters/subawardID"
      - name: invoiceID
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      tags: [invoices]
      summary: Upload an invoice PDF
      operationId: uploadInvoiceDocument
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "201":
          description: Created document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"

  # ── Statements of Work ─────────────────────────────────
  /grants/{grantID}/subawards/{subawardID}/sow:
    parameters:
      - $ref: "#/components/parameters/grantID"
      - $ref: "#/components/parameters/subawardID"
    get:
      tags: [sow]
      summary: List statements of work
      operationId: listStatementsOfWork
      responses:
        "200":
          description: Array of SOWs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StatementOfWork"
    post:
      tags: [sow]
      summary: Create a statement of work
      operationId: createStatementOfWork
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StatementOfWork"
      responses:
        "201":
          description: Created SOW
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatementOfWork"
  /grants/{grantID}/subawards/{subawardID}/sow/{sowID}:
    parameters:
      - $ref: "#/components/parameters/grantID"
      - $ref: "#/components/parameters/subawardID"
      - name: sowID
        in: path
        required: true
        schema:
          type: string
          format: uuid
    put:
      tags: [sow]
      summary: Update a statement of work
      operationId: updateStatementOfWork
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StatementOfWork"
      responses:
        "200":
          description: Updated SOW
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatementOfWork"
  /grants/{grantID}/subawards/{subawardID}/sow/{sowID}/upload-signed:
    parameters:
      - $ref: "#/components/parameters/grantID"
      - $ref: "#/components/parameters/subawardID"
      - name: sowID
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      tags: [sow]
      summary: Upload a signed SOW document
      operationId: uploadSignedSOW
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "200":
          description: Updated SOW with signed_doc_id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatementOfWork"

  # ── SOW Config ──────────────────────────────────────────
  /grants/{grantID}/sow-config:
    parameters:
      - $ref: "#/components/parameters/grantID"
    get:
      tags: [sow]
      summary: Get SOW document template configuration
      operationId: getSOWConfig
      responses:
        "200":
          description: SOW config for this grant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SOWConfig"
    put:
      tags: [sow]
      summary: Create or update SOW config
      operationId: upsertSOWConfig
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SOWConfig"
      responses:
        "200":
          description: Saved SOW config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SOWConfig"

  # ── SOW Render ──────────────────────────────────────────
  /grants/{grantID}/subawards/{subawardID}/sow/{sowID}/render:
    parameters:
      - $ref: "#/components/parameters/grantID"
      - $ref: "#/components/parameters/subawardID"
      - name: sowID
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags: [sow]
      summary: Render SOW as HTML document
      operationId: renderSOW
      responses:
        "200":
          description: HTML SOW document
          content:
            text/html:
              schema:
                type: string

  # ── SOW Personnel Descriptions ──────────────────────────
  /grants/{grantID}/subawards/{subawardID}/sow/{sowID}/personnel-descriptions:
    parameters:
      - $ref: "#/components/parameters/grantID"
      - $ref: "#/components/parameters/subawardID"
      - name: sowID
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags: [sow]
      summary: List personnel descriptions for a SOW
      operationId: listSOWPersonnelDescriptions
      responses:
        "200":
          description: Array of personnel descriptions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SOWPersonnelDescription"
    put:
      tags: [sow]
      summary: Create or update a personnel description
      operationId: upsertSOWPersonnelDescription
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SOWPersonnelDescription"
      responses:
        "200":
          description: Saved personnel description
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SOWPersonnelDescription"
  /grants/{grantID}/subawards/{subawardID}/sow/{sowID}/personnel-descriptions/{descID}:
    parameters:
      - $ref: "#/components/parameters/grantID"
      - $ref: "#/components/parameters/subawardID"
      - name: sowID
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: descID
        in: path
        required: true
        schema:
          type: string
          format: uuid
    delete:
      tags: [sow]
      summary: Delete a personnel description
      operationId: deleteSOWPersonnelDescription
      responses:
        "204":
          description: Deleted

  # ── SOW Line Item Descriptions ──────────────────────────
  /grants/{grantID}/subawards/{subawardID}/sow/{sowID}/line-item-descriptions:
    parameters:
      - $ref: "#/components/parameters/grantID"
      - $ref: "#/components/parameters/subawardID"
      - name: sowID
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags: [sow]
      summary: List line item descriptions for a SOW
      operationId: listSOWLineItemDescriptions
      responses:
        "200":
          description: Array of line item descriptions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SOWLineItemDescription"
    put:
      tags: [sow]
      summary: Create or update a line item description
      operationId: upsertSOWLineItemDescription
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SOWLineItemDescription"
      responses:
        "200":
          description: Saved line item description
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SOWLineItemDescription"
  /grants/{grantID}/subawards/{subawardID}/sow/{sowID}/line-item-descriptions/{descID}:
    parameters:
      - $ref: "#/components/parameters/grantID"
      - $ref: "#/components/parameters/subawardID"
      - name: sowID
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: descID
        in: path
        required: true
        schema:
          type: string
          format: uuid
    delete:
      tags: [sow]
      summary: Delete a line item description
      operationId: deleteSOWLineItemDescription
      responses:
        "204":
          description: Deleted

  # ── Documents ───────────────────────────────────────────
  /documents/{docID}:
    parameters:
      - name: docID
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags: [documents]
      summary: Get document metadata
      operationId: getDocument
      responses:
        "200":
          description: Document metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
  /documents/{docID}/download:
    parameters:
      - name: docID
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags: [documents]
      summary: Download a document file
      operationId: downloadDocument
      responses:
        "200":
          description: File binary
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
  # ── Overhead Rates ──────────────────────────────────────
...
  /institution-rates/{entityType}/{entityID}/fringe-rates/{fringeRateID}:
    parameters:
      - $ref: "#/components/parameters/entityType"
      - $ref: "#/components/parameters/entityID"
      - name: fringeRateID
        in: path
        required: true
        schema:
          type: string
          format: uuid
    delete:
      tags: [institution-rates]
      summary: Delete a fringe rate
      operationId: deleteFringeRate
      responses:
        "204":
          description: Deleted

  # ── Institution Budgets ─────────────────────────────────
...
  /institution-rates/{entityType}/{entityID}/budgets/{budgetID}:
    parameters:
      - $ref: "#/components/parameters/entityType"
      - $ref: "#/components/parameters/entityID"
      - $ref: "#/components/parameters/budgetID"
    delete:
      tags: [institution-rates]
      summary: Delete a budget
      operationId: deleteInstitutionBudget
      responses:
        "204":
          description: Deleted
...